breaking: organize factory function in one
NOTE:
  createI18n({ ... }) -> createI18n({ legacy: true, ... })
  createI18nComposer({ ... }) -> createI18n({ ... })
