+++ src/index.ts
@@ -1,4 +1,3 @@
-export const VERSION = __VERSION__
 export { Path, PathValue } from './path'
 export {
   Locale,
@@ -8,13 +7,7 @@ export {
   PostTranslationHandler
 } from './runtime'
 export * from './runtime/types'
-export {
-  MissingHandler,
-  ComposerOptions,
-  Composer,
-  createComposer
-} from './composer'
-export { useI18n } from './use'
+export { MissingHandler, ComposerOptions, Composer } from './composer'
 export {
   TranslateResult,
   Choice,
@@ -25,6 +18,7 @@ export {
   NumberFormatResult,
   Formatter,
   VueI18nOptions,
-  VueI18n,
-  createI18n
+  VueI18n
 } from './legacy'
+export { createI18n, useI18n } from './i18n'
+export const VERSION = __VERSION__
