+++ src/mixin.ts
@@ -4,7 +4,7 @@ import { Locale } from './runtime/context'
 import { Composer } from './composer'
 import {
   VueI18n,
-  createI18n,
+  createVueI18n,
   VueI18nOptions,
   TranslateResult,
   DateTimeFormatResult,
@@ -57,9 +57,9 @@ export function getMixin(
           optionsI18n.__i18n = options.__i18n
         }
         optionsI18n._root = composer
-        this.$i18n = createI18n(optionsI18n)
+        this.$i18n = createVueI18n(optionsI18n)
       } else if (options.__i18n) {
-        this.$i18n = createI18n({ __i18n: options.__i18n, _root: composer })
+        this.$i18n = createVueI18n({ __i18n: options.__i18n, _root: composer })
       } else if (this.$root && this.$root.proxy) {
         // root i18n
         // TODO: should resolve type inference
